Clean up system role modals and add usage tips.
